requirements.txt
streamlit
pydeck
plotly
pandas
numpy
requests
asyncio-mqtt
python-dotenv
langchain
langsmith
langgraph
google-generativeai